Trivial cut-paste changes to worker and worker_state_machine
